requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pymilvus==2.6.6
python-dotenv==1.0.1
langchain>=0.1.0
google-genai==0.3.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
streamlit>=1.55.0